refactor(snmptraps): migrate forwarder component to v2
