+++ comp/snmptraps/forwarder/def/component.go
