+++ comp/snmptraps/forwarder/impl/forwarder_test.go
@@ -18,7 +18,8 @@ import (
 	configfx "github.com/DataDog/datadog-agent/comp/snmptraps/config/fx"
 	formatter "github.com/DataDog/datadog-agent/comp/snmptraps/formatter/def"
 	formatterfx "github.com/DataDog/datadog-agent/comp/snmptraps/formatter/fx"
-	"github.com/DataDog/datadog-agent/comp/snmptraps/forwarder"
+	forwarder "github.com/DataDog/datadog-agent/comp/snmptraps/forwarder/def"
+	forwarderfx "github.com/DataDog/datadog-agent/comp/snmptraps/forwarder/fx"
 	listener "github.com/DataDog/datadog-agent/comp/snmptraps/listener/def"
 	listenerfx "github.com/DataDog/datadog-agent/comp/snmptraps/listener/fx"
 	"github.com/DataDog/datadog-agent/comp/snmptraps/packet"
@@ -44,7 +45,7 @@ func setUp(t *testing.T) *services {
 		senderhelper.Opts,
 		formatterfx.MockModule(),
 		listenerfx.MockModule(),
-		Module(),
+		forwarderfx.Module(),
 	)
 	return &s
 }
