+++ comp/snmptraps/server/impl/server.go
@@ -23,8 +23,8 @@ import (
 	configfx "github.com/DataDog/datadog-agent/comp/snmptraps/config/fx"
 	formatter "github.com/DataDog/datadog-agent/comp/snmptraps/formatter/def"
 	formatterimpl "github.com/DataDog/datadog-agent/comp/snmptraps/formatter/impl"
-	"github.com/DataDog/datadog-agent/comp/snmptraps/forwarder"
-	"github.com/DataDog/datadog-agent/comp/snmptraps/forwarder/forwarderimpl"
+	forwarder "github.com/DataDog/datadog-agent/comp/snmptraps/forwarder/def"
+	forwarderimpl "github.com/DataDog/datadog-agent/comp/snmptraps/forwarder/fx"
 	listener "github.com/DataDog/datadog-agent/comp/snmptraps/listener/def"
 	listenerfx "github.com/DataDog/datadog-agent/comp/snmptraps/listener/fx"
 	oidresolverfx "github.com/DataDog/datadog-agent/comp/snmptraps/oidresolver/fx"
